[#127] CalendarView의 View 관련 함수를 View Extension으로 이동
