+++ EUMMEYO/Extension/View+.swift
@@ -1,0 +1,40 @@
+//
+//  View+.swift
+//  EUMMEYO
+//
+//  Created by 김동현 on 5/4/25.
+//
+
+import SwiftUI
+
+// MARK: - UI Design Heplher functions
+extension View {
+    // 부모 View의 가로 공간을 최대한 차지하도록 설정. -> 왼쪽, 오른쪽, 가운데 정렬로 배치
+    func hLeading() -> some View {
+        self
+            .frame(maxWidth: .infinity, alignment: .leading)
+    }
+    
+    func hTrailing() -> some View {
+        self
+            .frame(maxWidth: .infinity, alignment: .trailing)
+    }
+    
+    func hCenter() -> some View {
+        self
+            .frame(maxWidth: .infinity, alignment: .center)
+    }
+    
+    // MARK: - Safe Area
+    func getSafeArea() -> UIEdgeInsets {
+        guard let screen = UIApplication.shared.connectedScenes.first as? UIWindowScene else {
+            return .zero
+        }
+        
+        guard let safeArea = screen.windows.first?.safeAreaInsets else {
+            return .zero
+        }
+        
+        return safeArea
+    }
+}
